Initial Q2, need to improve error_est
Have a working interpolant and plotting of the results, error_estimate runs but gives 1e-16 which seems unreasonably good. Probably bc spline goes through the points *exactly* so I think I need to use the derivative information to either compare to the spline derivative or to get a temperature estimate away from a data point and compare to that.